Rename create post action to uploadImage
